Add embedding modules and extend docs - Add `semantic_search.md` guide - Extend existing docs - Move implementations out of `Torus` module
