+++ config/config.exs
@@ -10,7 +10,15 @@ config :torus, Torus.Test.Repo,
   priv: "test/support",
   show_sensitive_data_on_connection_error: true,
   stacktrace: true,
-  username: "root",
+  # Will this work for CI?
+  # username: "root",
   url: System.get_env("POSTGRES_URL") || "postgres://localhost:5432/torus_test"
 
 config :torus, ecto_repos: [Torus.Test.Repo]
+
+config :torus, batcher: Torus.Embeddings.Batcher
+
+config :torus, Torus.Embeddings.Batcher,
+  max_batch_size: 10,
+  default_batch_timeout: 100,
+  embedding_module: Torus.Embeddings.HuggingFace
