+++ lib/torus/embeddings/hugging_face.ex
@@ -1,10 +1,41 @@
 if Code.ensure_loaded?(Req) do
   defmodule Torus.Embeddings.HuggingFace do
     @moduledoc """
-    Embedding generator using Hugging Face Inference API.
+    A wrapper around Hugging Face API. It allows you to generate embeddings using a variety of models available on Hugging Face.
 
-    This module requires the `req` library to be added to your dependencies.
+    To use it:
+
+    - Add the following to your `config.exs`:
+
+      ```elixir
+      config :torus, embedding_module: Torus.Embeddings.HuggingFace
+      ```
+
+    - Add `req` to your `mix.exs` dependencies:
+
+      ```elixir
+      def deps do
+      [
+        {:req, "~> 0.5"}
+      ]
+      end
+      ```
+
+    - Add an API token for hugging face to your `runtime.exs`. You can get your token [here](https://huggingface.co/settings/tokens).
+
+      ```elixir
+      config :torus, Torus.Embeddings.HuggingFace, token: System.get_env("HUGGING_FACE_API_KEY")
+      ```
+
+    By default, it uses `sentence-transformers/all-MiniLM-L6-v2` model, but you can specify a different model by explicitly passing `model` to the config:
+
+    ```elixir
+    config :torus, Torus.Embeddings.HuggingFace, model: "your/model"
+    ```
+
+    See `Torus.semantic/5` on how to use this module to introduce semantic search in your application.
     """
+    require Torus
     alias Torus.Embeddings.Common
 
     @behaviour Torus.Embedding
