+++ lib/torus/search/pattern_match.ex
@@ -1,0 +1,57 @@
+defmodule Torus.Search.PatternMatch do
+  @moduledoc false
+
+  def ilike(query, bindings, qualifiers, term) do
+    qualifiers = List.wrap(qualifiers)
+
+    where_ast =
+      Enum.reduce(qualifiers, false, fn qualifier, conditions_acc ->
+        quote do
+          dynamic(
+            [unquote_splicing(bindings)],
+            ilike(unquote(qualifier), ^unquote(term)) or ^unquote(conditions_acc)
+          )
+        end
+      end)
+
+    quote do
+      where(unquote(query), ^unquote(where_ast))
+    end
+  end
+
+  def like(query, bindings, qualifiers, term) do
+    qualifiers = List.wrap(qualifiers)
+
+    where_ast =
+      Enum.reduce(qualifiers, false, fn qualifier, conditions_acc ->
+        quote do
+          dynamic(
+            [unquote_splicing(bindings)],
+            like(unquote(qualifier), ^unquote(term)) or ^unquote(conditions_acc)
+          )
+        end
+      end)
+
+    quote do
+      where(unquote(query), ^unquote(where_ast))
+    end
+  end
+
+  def similar_to(query, bindings, qualifiers, term) do
+    qualifiers = List.wrap(qualifiers)
+
+    where_ast =
+      Enum.reduce(qualifiers, false, fn qualifier, conditions_acc ->
+        quote do
+          dynamic(
+            [unquote_splicing(bindings)],
+            operator(unquote(qualifier), "SIMILAR TO", ^unquote(term)) or ^unquote(conditions_acc)
+          )
+        end
+      end)
+
+    quote do
+      where(unquote(query), ^unquote(where_ast))
+    end
+  end
+end
