+++ mix.exs
@@ -37,6 +37,7 @@ defmodule Torus.MixProject do
         source_ref: "v#{@version}",
         source_url: @source_url,
         extra_section: "GUIDES",
+        groups_for_modules: groups_for_modules(),
         formatters: ["html"],
         extras: extras(),
         skip_undefined_reference_warnings_on: ["CHANGELOG.md"]
@@ -53,6 +54,19 @@ defmodule Torus.MixProject do
 
   defp elixirc_paths(:test), do: ["lib", "test/support"]
   defp elixirc_paths(_env), do: ["lib"]
+
+  defp groups_for_modules do
+    [
+      Embeddings: [
+        Torus.Embedding,
+        Torus.Embeddings.LocalNxServing,
+        Torus.Embeddings.OpenAI,
+        Torus.Embeddings.HuggingFace,
+        Torus.Embeddings.PostgresML,
+        Torus.Embeddings.Batcher
+      ]
+    ]
+  end
 
   defp deps do
     [
@@ -103,6 +117,8 @@ defmodule Torus.MixProject do
 
   defp extras do
     [
+      "guides/semantic_search.md",
+
       # TODO Add more guides
       "CHANGELOG.md": [title: "Changelog"]
     ]
